% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_density_calculation.R
\name{snotel_density_calculation}
\alias{snotel_density_calculation}
\title{snotel_density_calculation}
\usage{
snotel_density_calculation(df, swe, snow_depth)
}
\arguments{
\item{df}{dataframe}
}
\value{
column containing calculated density values
}
\description{
This function calculates density from SWE and snow depth.
}
\examples{
\dontrun{
# Load your dataframe
data(snotel_000)

# Calculate density
snotel_000 <- snotel_000 \%>\%
  snotel_density_calculation()
}

}
