% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_derivative.R
\name{snotel_derivative}
\alias{snotel_derivative}
\title{snotel_derivative}
\usage{
snotel_derivative(df, col_name)
}
\arguments{
\item{df}{dataframe}

\item{input_column}{Column from which the derivative is being computed from}
}
\value{
Column with computed derivative
}
\description{
This function computes the derivative of a specified column.
}
\examples{
\dontrun{
# Load dataframe
data(snotel_000)

# Calculate derivative for 'soilm_p2' column
snotel_000 <- snotel_000 \%>\%
  snotel_derivative(soilm_p2)
}
}
