% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_find_smpulse.R
\name{snotel_detect_smpulse}
\alias{snotel_detect_smpulse}
\title{snotel_detect_smpulse}
\usage{
snotel_detect_smpulse(
  df,
  soil_moisture_col,
  doy_col,
  year_col,
  swe_col,
  min_swe_window_before,
  max_swe_window_after,
  variability_multiplier,
  derivative_length
)
}
\arguments{
\item{df}{input processed SNOTEL df}

\item{soil_moisture_col}{column containing soil moisture of desired depth}

\item{doy_col}{column containing DOY}

\item{year_col}{column containing year}

\item{swe_col}{column containing SWE}

\item{max_swe_window_after}{define number of days after peak SWE that the algorithm should search for a pulse}

\item{variability_multiplier}{define standard deviation multiplier (higher multipliers will look for bigger pulses, lower multipliers will look for smaller pulses)}

\item{derivative_length}{define forward-looking derivative length (longer length will smooth the derivaive more, while shorter length will smooth the derivative less)}

\item{min_swe_mindow_before}{define number of days before peak SWE that the algorithm should search for a pulse}
}
\value{
dataframe with DOY of soil moisture pulse for all available years
}
\description{
Soil moisture pulse detection from daily SNOTEL metrics of soil moisture. 
The soil moisture pulse serves as a proxy for melt output from the snowpack.
}
\examples{
\dontrun{

 # Find pulses for a given SNOTEL processed df:
 # Change params as needed!
pulse_results <- snotel_detect_smpulse(
df = df, 
soil_moisture_col = "soilm_p2", 
doy_col = "doy", 
year_col = "year", 
swe_col = "swe", 
min_swe_window_before = 70, 
max_swe_window_after = 30,
variability_multiplier = 1.25, 
derivative_length = 2)

}

}
