% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_downloader_daily.R
\name{snotel_downloader_daily}
\alias{snotel_downloader_daily}
\title{snotel_downloader_daily}
\usage{
snotel_downloader_daily(site, state, start_date, end_date, save_dir)
}
\arguments{
\item{site}{SNOTEL numeric site code}

\item{state}{State where SNOTEL station is located in}

\item{start_date}{Start date, in YYYY-MM-DD format}

\item{end_date}{End date, in YYYY-MM-DD format}

\item{save_dir}{Path to directory where files will be saved}
}
\value{
CSV in desired directory with SNOTEL station data
}
\description{
This function downloads daily SNOTEL data from one (or multiple sites) by
accessing the NRCS report generator.
}
\examples{
\dontrun{

# Define directory to save output CSV files:
snotel_dir = "path/to/directory"

# Downloading data for one site, UT-907 for example:
snotel_downloader_daily(site = "907", state = "UT", start_date = "2013-01-01", 
end_date = "2014-08-01", save_dir = snotel_dir)

# Downloading data for multiple sites:

# Define sites in a tibble:
df <- tibble(site = c(531, 335, 505, 485, 802),
state = c("CO" ,"CO", "CO", "CO", "CO"))

# Using walk2, from purrr package for a tibble/df of sites:
walk2(.x = df$site, .y = df$state, ~snotel_downloader(site = .x, state=.y, 
start_date = "2013-01-01", end_date = "2015-12-31", save_dir = snotel_dir))

}
}
