% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_medfilter.R
\name{snotel_medfilter}
\alias{snotel_medfilter}
\title{snotel_medfilter}
\usage{
snotel_medfilter(x, n)
}
\arguments{
\item{x}{column in specified dataframe}

\item{n}{median filtering factor (a number)}
}
\value{
median filtered column in specified dataframe
}
\description{
This function runs a median filter on a column in a dataframe.
}
\examples{
\dontrun{
# Load your dataframe
data(snotel_000)

# Run median filter on specified column, here we are choosing density with n = 3:
CO_335_sample \%>\%
  mutate(density_medfilter = snotel_medfilter(density_calc, n = 3))

}

}
