% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_metadata.R
\name{snotel_metadata}
\alias{snotel_metadata}
\title{snotel_metadata}
\usage{
snotel_metadata(site_codes = NULL, states = NULL)
}
\arguments{
\item{site_codes}{Numeric code(s) for one or multiple SNOTEL stations}

\item{states}{Two-letter state code}
}
\value{
dataframe with metadata for desired SNOTEL sites

#' @examples
\dontrun{
# Select by site code:
site_1267 <- snotel_metadata(site_codes = c("1267"))

# Select by state:
state_co <- snotel_metadata(states = c("CO"))

# Select full metadata file for all sites/states:
full_metadata <- snotel_metadata()
}
}
\description{
Get metadata for SNOTEL stations in the western U.S.
}
