% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_metric.R
\name{snotel_metric}
\alias{snotel_metric}
\title{snotel_metric}
\usage{
snotel_metric(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
Dataframe with metric values in SWE, snow depth, precip, and air temp
}
\description{
This function converts values from imperial to metric system.
}
