% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_rename_col_daily.R
\name{snotel_rename_col_daily}
\alias{snotel_rename_col_daily}
\title{snotel_rename_col_daily}
\usage{
snotel_rename_col_daily(df, rename_map = NULL)
}
\arguments{
\item{df}{Input SNOTEL dataframe}

\item{rename_map}{named vector where the names are the new column names and 
the values are the existing column names. If custom rename_map is desired,
arrange it in "newcolumn" = "oldcolumn" format.}
}
\value{
Dataframe with renamed columns
}
\description{
This function renames the columns in the raw SNOTEL CSV download so that
they are more friendly for R and easier to understand. This function also
creates new columns for various date/time metrics that will be required
later on in general scripting.
}
