% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_medfilter.R
\name{snotel_medfilter}
\alias{snotel_medfilter}
\title{snotel_medfilter}
\usage{
snotel_medfilter(x, n)
}
\arguments{
\item{x}{column in specified dataframe}

\item{n}{median filtering factor (a number)}
}
\value{
median filtered column in specified dataframe

#' @examples
\dontrun{
# Load your dataframe
data(snotel_000)

# Run median filter on specified column, here we are choosing density with n = 3:
snotel_000 <- snotel_medfilter(density, 3)

}
}
\description{
This function runs a median filter on a column in a dataframe.
}
